pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = ["sitnarf <sitnarf@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pandas = "^2.2.3"
pytest = "^8.3.3"
streamlit = "^1.38.0"
scikit-surprise = "^1.1.4"
mlflow = "^2.16.2"
prefect = "^3.0.3"
numpy = "1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
